Move Postgres adapter to separate NPM package and namespace
