+++ package.json
@@ -14,6 +14,7 @@
     "skipruntime-ts/tests",
     "skipruntime-ts/server",
     "skipruntime-ts/examples",
+    "skipruntime-ts/adapters/postgres",
     "sql/ts",
     "sql/ts/tests",
     "examples/hackernews/reactive_service"
