+++ skipruntime-ts/adapters/postgres/package.json
@@ -1,11 +1,9 @@
 {
-  "name": "@skipruntime/helpers",
-  "version": "0.0.8",
+  "name": "@skip-adapter/postgres",
+  "version": "0.0.2",
   "type": "module",
   "exports": {
-    ".": "./dist/index.js",
-    "./rest.js": "./dist/rest.js",
-    "./external.js": "./dist/external.js"
+    ".": "./dist/index.js"
   },
   "scripts": {
     "build": "tsc",
@@ -14,16 +12,17 @@
   },
   "author": "SkipLabs",
   "license": "MIT",
-  "description": "",
+  "description": "Adapter providing reactive access to PostgreSQL from the Skip runtime",
   "engines": {
     "node": ">=22.6.0 <23.0.0"
   },
   "dependencies": {
-    "eventsource": "^2.0.2",
-    "express": "^4.21.1",
+    "pg": "^8.13.1",
+    "pg-format": "^1.0.4",
     "@skipruntime/core": "0.0.5"
   },
   "devDependencies": {
-    "@types/eventsource": "^1.1.15"
+    "@types/pg": "^8.11.10",
+    "@types/pg-format": "^1.0.5"
   }
 }
