+++ skipruntime-ts/adapters/postgres/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@skiplabs/tsconfig"
+}
