+++ skipruntime-ts/helpers/src/index.ts
@@ -7,7 +7,6 @@
 export {
   type ExternalResource,
   GenericExternalService,
-  PostgresExternalService,
   Polled,
 } from "./external.js";
 export { SkipExternalService } from "./remote.js";
