+++ skipruntime-ts/tests/package.json
@@ -9,14 +9,17 @@
     "test": "LD_LIBRARY_PATH=$(realpath ../../build/skipruntime) mocha"
   },
   "devDependencies": {
-    "@types/mocha": "^10.0.10"
+    "@types/mocha": "^10.0.10",
+    "@types/pg": "^8.11.10"
   },
   "dependencies": {
     "earl": "^1.3.0",
     "mocha": "^11.0.1",
+    "pg": "^8.13.1",
     "@skipruntime/core": "0.0.5",
     "@skipruntime/helpers": "0.0.8",
     "@skipruntime/native": "0.0.4",
-    "@skipruntime/wasm": "0.0.6"
+    "@skipruntime/wasm": "0.0.6",
+    "@skip-adapter/postgres": "0.0.2"
   }
 }
